+++ test/helpers.js
@@ -699,3 +699,10 @@ export function setPasswordlessFlowType(contactMethod, flowType) {
 export function isReact16() {
     return process.env.IS_REACT_16 === "true";
 }
+
+export async function getResetPasswordFormBackButton(page) {
+    const backButtonSelector =
+        "[data-supertokens='headerTitle resetPasswordHeaderTitle'] > [data-supertokens='backButton backButtonCommon']";
+
+    return await waitForSTElement(page, backButtonSelector);
+}